Cover the various check access helpers in the fileio unit tests.
Please note that we effectively run unit tests as uid 0 when using the
containerized setup, and that may yield rather surprising results when testing
file access, because super-users have access to everything despite what the
permissions indicate.

The added tests take uid 0 into account and will succeed, but we should really
eliminate these privileges when unit testing as they can easily cause a lot of
head-scratching.

The tests also uncovered a couple of additional corner case bugs in the access
checks, which don't show up with the uid 0 case but need to be addressed for
the general benefit. They have been registered in issue #378 for further work.