Force Show Peers date to display on same format as the date picker by wrapping (#385)
it in a dummy input field without border.
Limit date picker UI and backend accepted date range to dates between 7 days
from today and ~10 years in the future, with fallback to default 365 days if
an invalid value is given. Highlight the fallback in output.
The end date value is automatically interpreted as until midnight and the
defaults are now read from mig/shared/defaults.py both for UI and backend.

Signed-off-by: Jonas Bardino <bardino@science.ku.dk>